+++ src/com/amplitude/api/AmplitudeClient.java
@@ -17,7 +17,6 @@ import com.squareup.okhttp.Response;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.w3c.dom.Text;
 
 import android.app.Application;
 import android.content.Context;